Fix parsing generics within raises postfix statement
