+++ src/parser/file.rs
@@ -1,14 +1,12 @@
 use crate::lexer::token::Token;
 use crate::parser::_type::parse_conditions;
-use crate::parser::_type::parse_generics;
 use crate::parser::_type::parse_id;
 use crate::parser::_type::parse_type;
 use crate::parser::ast::ASTNode;
 use crate::parser::ast::ASTNodePos;
 use crate::parser::block::parse_block;
 use crate::parser::block::parse_statements;
-use crate::parser::definition::parse_fun_arg;
-use crate::parser::expression::parse_expression;
+use crate::parser::class::parse_class;
 use crate::parser::iterator::TPIterator;
 use crate::parser::parse_result::ParseErr::*;
 use crate::parser::parse_result::ParseResult;
@@ -35,7 +33,7 @@ pub fn parse_import(it: &mut TPIterator) -> ParseResult {
         it.eat_if(Token::Comma);
         Ok(())
     })?;
-    let _as = it.parse_vec_if_token(Token::As, &parse_as, "import")?;
+    let _as = it.parse_vec_if(Token::As, &parse_as, "import")?;
 
     let (en_line, en_pos) = match (import.last(), _as.last()) {
         (_, Some(token_pos)) => (token_pos.en_line, token_pos.en_pos),
@@ -62,73 +60,6 @@ fn parse_as(it: &mut TPIterator) -> ParseResult<Vec<ASTNodePos>> {
     })?;
 
     Ok(aliases)
-}
-
-pub fn parse_class(it: &mut TPIterator) -> ParseResult {
-    let (st_line, st_pos) = it.start_pos()?;
-    it.eat(Token::Class, "class")?;
-    let _type = it.parse(&parse_type, "name")?;
-
-    let mut args = vec![];
-    it.peek_while_not_token(Token::LRBrack, &mut |it, token_pos, no| match token_pos.token {
-        Token::Def => {
-            args.push(*it.parse(&parse_fun_arg, format!("constructor argument {}", no).as_str())?);
-            it.eat_if(Token::Comma);
-            Ok(())
-        }
-        _ => Err(TokenErr {
-            expected: Token::Def,
-            actual:   token_pos.clone(),
-            message:  String::from("class")
-        })
-    })?;
-
-    let parents = it.parse_vec_if_token(
-        Token::IsA,
-        &|it| {
-            let mut parents = vec![];
-            it.peek_while_not_token(Token::NL, &mut |it, token_pos, no| match token_pos.token {
-                Token::Id(_) => {
-                    parents.push(*it.parse(&parse_parent, format!("parent {}", no).as_str())?);
-                    it.eat_if(Token::Comma);
-                    Ok(())
-                }
-                _ => Err(TokenErr {
-                    expected: Token::Id(String::new()),
-                    actual:   token_pos.clone(),
-                    message:  format!("parent {}", no)
-                })
-            })?;
-            Ok(parents)
-        },
-        "parents"
-    )?;
-
-    let body = it.parse(&parse_block, "class body")?;
-    let (en_line, en_pos) = (body.en_line, body.en_pos);
-    let node = ASTNode::Class { _type, args, parents, body };
-    Ok(Box::from(ASTNodePos { st_line, st_pos, en_line, en_pos, node }))
-}
-
-pub fn parse_parent(it: &mut TPIterator) -> ParseResult {
-    let (st_line, st_pos) = it.start_pos()?;
-
-    let id = it.parse(&parse_id, "parent id")?;
-    let generics = it.parse_vec_if_token(Token::LSBrack, &parse_generics, "parent generics")?;
-    let mut args = vec![];
-    it.peek_while_not_token(Token::LRBrack, &mut |it, _, no| {
-        args.push(*it.parse(&parse_expression, format!("parent argument {}", no).as_str())?);
-        it.eat_if(Token::Comma);
-        Ok(())
-    })?;
-
-    let (en_line, en_pos) = match (generics.last(), args.last()) {
-        (_, Some(tp)) => (tp.en_line, tp.en_pos),
-        (Some(tp), _) => (tp.en_line, tp.en_pos),
-        _ => (id.en_line, id.en_pos)
-    };
-    let node = ASTNode::Parent { id, generics, args };
-    Ok(Box::from(ASTNodePos { st_line, st_pos, en_line, en_pos, node }))
 }
 
 pub fn parse_script(it: &mut TPIterator) -> ParseResult {
@@ -204,7 +135,7 @@ pub fn parse_type_def(it: &mut TPIterator) -> ParseResult {
                 it.eat(Token::IsA, "type definition")?;
                 let _type = it.parse(&parse_type, "type definition")?;
                 let conditions =
-                    it.parse_vec_if_token(Token::When, &parse_conditions, "type definitions")?;
+                    it.parse_vec_if(Token::When, &parse_conditions, "type definition")?;
 
                 let (en_line, en_pos) = (_type.en_line, _type.en_pos);
                 let node = ASTNode::TypeAlias { _type: _type.clone(), conditions };
